# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Skolara API",
    "description": "Backend for searching colleges, managing applications, and processing payments.",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:8000/api" },
    { "url": "https://devapi.skolara.in/api" }
  ],
...
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/colleges": {
      "get": {
        "tags": ["Colleges"],
        "summary": "List colleges",
        "operationId": "getColleges",
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "schema": { "type": "number" },
            "required": false,
            "description": "Filter by category (college_type_id)"
          },
          {
            "in": "query",
            "name": "search",
            "schema": { "type": "string" },
            "required": false,
            "description": "Search by name (prefix match) or short name (contains)"
          },
          {
            "in": "query",
            "name": "name",
            "schema": { "type": "string" },
            "required": false,
            "description": "Alias of `search`. If both provided, `search` takes precedence."
          }
        ],
        "responses": {
          "200": {
            "description": "List of colleges",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollegeListResponse"
                },
                "examples": {
                  "ok": {
                    "value": [
                      {
                        "id": "clg_abc123",
                        "name": "Skolara Institute of Tech",
                        "area": "Koramangala",
                        "city": "Bengaluru",
                        "logo_url": "https://cdn.example.com/logo.png",
                        "is_applied": false
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameter (e.g., category not a number)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "invalidCategory": {
                    "value": { "error": "Invalid 'category' value" }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/JWTUnauthorized" }
        }
      }
    },
    "/colleges/{id}": {
      "get": {
        "tags": ["Colleges"],
        "summary": "Get college by ID",
        "operationId": "getCollegeById",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "College details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CollegeDetail" }
              }
            }
          },
          "401": { "$ref": "#/components/responses/JWTUnauthorized" },
          "404": {
            "description": "College not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "nf": { "value": { "error": "College not found" } }
                }
              }
            }
          }
        }
      }
    },
...
    "/colleges/save": {
      "post": {
        "tags": ["Colleges"],
        "summary": "Toggle save college",
        "operationId": "saveCollege",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SaveCollegeRequest" },
              "examples": {
                "default": { "value": { "college_id": "clg_abc123" } }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "College saved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveCollegeResponse"
                },
                "examples": {
                  "saved": {
                    "value": { "message": "College saved", "saved": true }
                  }
                }
              }
            }
          },
          "200": {
            "description": "Save toggled (either saved or removed)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SaveCollegeResponse"
                },
                "examples": {
                  "removed": {
                    "value": {
                      "message": "College removed from saved",
                      "saved": false
                    }
                  },
                  "reSaved": {
                    "value": { "message": "College saved", "saved": true }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/JWTUnauthorized" },
          "404": {
            "description": "College not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "nf": { "value": { "error": "College not found" } }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Colleges"],
        "summary": "List saved colleges",
        "operationId": "getSavedColleges",
        "responses": {
          "200": {
            "description": "Saved colleges for the authenticated user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CollegeListResponse"
                },
                "examples": {
                  "ok": {
                    "value": [
                      {
                        "id": "clg_abc123",
                        "name": "Skolara Institute of Tech",
                        "area": "Koramangala",
                        "city": "Bengaluru",
                        "logo_url": "https://cdn.example.com/logo.png",
                        "is_applied": true
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/JWTUnauthorized" }
        }
      }
    },
    "/colleges/applied": {
      "get": {
        "tags": ["Colleges"],
        "summary": "List applied colleges",
        "operationId": "getAppliedColleges",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of colleges the user has applied to",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AppliedCollegeListResponse"
                },
                "examples": {
                  "ok": {
                    "value": [
                      {
                        "id": "app_abc123",
                        "applied_at": "2025-03-01T10:15:00.000Z",
                        "ready_to_pay": false,
                        "documents_verified_at": null,
                        "reviewed_at": null,
                        "submitted_at": null,
                        "accepted_at": null,
                        "paid_at": null,
                        "paid": false,
                        "is_active": true,
                        "amount": 5000,
                        "application_status_id": 2,
                        "status": "Pending",
                        "hsc_path": null,
                        "hsc_signed_url": null,
                        "sslc_path": "docs/sslc/abc.pdf",
                        "sslc_signed_url": "https://signed-url.com/docs/sslc/abc.pdf",
                        "college": {
                          "id": "clg_abc123",
                          "name": "Skolara Institute of Tech",
                          "area": "Koramangala",
                          "city": "Bengaluru",
                          "logo_url": "https://cdn.example.com/logo.png"
                        }
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/JWTUnauthorized" }
        }
      }
    },
    "/colleges/apply": {
      "post": {
        "tags": ["Colleges"],
        "summary": "Apply to a college",
        "operationId": "applyCollege",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyCollegeRequest"
              },
              "examples": {
                "default": {
                  "value": {
                    "college_id": "clg_abc123",
                    "hsc_path": "docs/hsc/123.pdf",
                    "sslc_path": "docs/sslc/123.pdf"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "College application created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplyCollegeResponse"
                },
                "examples": {
                  "created": {
                    "value": {
                      "message": "College applied",
                      "applicationId": "app_abc123"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request (e.g., already applied for the college)",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "alreadyApplied": {
                    "value": {
                      "message": "You have already applied for this college"
                    }
                  }
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/JWTUnauthorized" },
          "404": {
            "description": "College not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" },
                "examples": {
                  "notFound": {
                    "value": { "error": "College not found" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Standard JWT bearer token (used on protected endpoints)."
      }
    },
    "responses": {
      "JWTUnauthorized": {
        "description": "Missing, malformed, expired, or invalid JWT",
        "headers": {
          "WWW-Authenticate": {
            "schema": { "type": "string" },
            "description": "Auth challenge header (e.g., Bearer realm and error details)"
          }
        },
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/JWTErrorResponse" },
            "examples": {
              "missingHeader": {
                "value": {
                  "error": "UNAUTHORIZED",
                  "message": "Missing or malformed Authorization header",
                  "statusCode": 401
                }
              },
              "invalid": {
                "value": {
                  "error": "INVALID_TOKEN",
                  "message": "JWT signature verification failed",
                  "statusCode": 401
                }
              },
              "expired": {
                "value": {
                  "error": "TOKEN_EXPIRED",
                  "message": "JWT access token has expired",
                  "statusCode": 401
                }
              }
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        },
        "required": ["error"]
      },
...
      "JWTErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "UNAUTHORIZED",
              "INVALID_TOKEN",
              "TOKEN_EXPIRED",
              "FORBIDDEN"
            ],
            "description": "JWT validation error type"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error details"
          },
          "statusCode": {
            "type": "integer",
            "example": 401,
            "description": "HTTP status code"
          }
        },
        "required": ["error"]
      },
...
      "CollegeListItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "area": { "type": "string", "nullable": true },
          "city": { "type": "string", "nullable": true },
          "logo_url": { "type": "string", "format": "uri", "nullable": true },
          "is_applied": { "type": "boolean" }
        },
        "required": ["id", "name", "is_applied"]
      },
      "CollegeListResponse": {
        "type": "array",
        "items": { "$ref": "#/components/schemas/CollegeListItem" }
      },
      "EligibilityItem": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "criteria_type": { "type": "string" },
          "criteria": { "type": "string" }
        },
        "required": ["id", "criteria_type", "criteria"]
      },
      "ScholarshipItem": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "schemes": { "type": "string" }
        },
        "required": ["id", "schemes"]
      },
      "CollegeDetail": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "area": { "type": "string", "nullable": true },
          "city": { "type": "string", "nullable": true },
          "logo_url": { "type": "string", "format": "uri", "nullable": true },
          "cover_url": { "type": "string", "format": "uri", "nullable": true },
          "is_saved": { "type": "boolean" },
          "is_applied": { "type": "boolean" },
          "description": { "type": "string", "nullable": true },
          "deadline": { "type": "string", "format": "date", "nullable": true },
          "eligibility": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/EligibilityItem" }
          },
          "scholarships": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ScholarshipItem" }
          }
        },
        "required": [
          "id",
          "name",
          "is_saved",
          "is_applied",
          "eligibility",
          "scholarships"
        ]
      },
...
      "SaveCollegeRequest": {
        "type": "object",
        "properties": {
          "college_id": { "type": "string" }
        },
        "required": ["college_id"],
        "additionalProperties": false
      },
      "SaveCollegeResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "saved": { "type": "boolean" }
        },
        "required": ["message", "saved"]
      },
      "AppliedCollegeItem": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "applied_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "ready_to_pay": { "type": "boolean" },
          "documents_verified_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "reviewed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "accepted_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "paid_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "paid": { "type": "boolean" },
          "is_active": { "type": "boolean" },
          "amount": { "type": "number", "nullable": true },
          "application_status_id": { "type": "number" },
          "status": { "type": "string" },
          "hsc_path": { "type": "string", "nullable": true },
          "hsc_signed_url": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "sslc_path": { "type": "string", "nullable": true },
          "sslc_signed_url": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "college": {
            "type": "object",
            "properties": {
              "id": { "type": "string" },
              "name": { "type": "string" },
              "area": { "type": "string", "nullable": true },
              "city": { "type": "string", "nullable": true },
              "logo_url": {
                "type": "string",
                "format": "uri",
                "nullable": true
              }
            },
            "required": ["id", "name"]
          }
        },
        "required": ["id", "status", "is_active", "paid", "college"]
      },
      "AppliedCollegeListResponse": {
        "type": "array",
        "items": { "$ref": "#/components/schemas/AppliedCollegeItem" }
      },
      "ApplyCollegeRequest": {
        "type": "object",
        "properties": {
          "college_id": { "type": "string", "format": "uuid" },
          "hsc_path": { "type": "string" },
          "sslc_path": { "type": "string" }
        },
        "required": ["college_id", "sslc_path"]
      },
      "ApplyCollegeResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string" },
          "applicationId": { "type": "string" }
        },
        "required": ["message", "applicationId"]
      },
...
    }
  }
}